+++ services/gateway/pkg/product/contracts/create_product_request.go
@@ -7,7 +7,7 @@ type CreateProductRequest struct {
 	Name          string  `json:"name" validate:"required,max=50"`
 	Description   *string `json:"description" validate:"omitempty,max=50"`
 	Sku           *string `json:"sku" validate:"omitempty,max=12"`
-	Quantity      uint64  `json:"quantity" validate:"required"`
+	Quantity      uint32  `json:"quantity" validate:"required"`
 	Price         uint64  `json:"price" validate:"required"`
 	ProductStatus uint32  `json:"productStatus" validate:"required"`
 }
