+++ services/gateway/pkg/product/contracts/get_product_response.go
@@ -6,7 +6,7 @@ type GetProductResponse struct {
 	Name          string   `json:"name"`
 	Description   *string  `json:"description"`
 	Sku           *string  `json:"sku"`
-	Quantity      uint64   `json:"quantity"`
+	Quantity      uint32   `json:"quantity"`
 	Price         uint64   `json:"price"`
 	Images        []string `json:"images"`
 	ProductStatus uint32   `json:"productStatus"`
